requirements/test.txt
coverage
flake8<5
pytest
pytest-cov
pytest-django
pytest-mock
pytest-deadfixtures
pytest-randomly
djangorestframework
django-mptt
flake8-bugbear
flake8-broken-line
flake8-commas
flake8-comprehensions
flake8-debugger
flake8-eradicate
flake8-isort
flake8-string-format
isort
